move onboarding complete to OnboardingController
