+++ apps/onboarding-ms/src/main/java/it/pagopa/selfcare/onboarding/controller/OnboardingController.java
@@ -15,8 +15,12 @@ import jakarta.validation.Valid;
 import jakarta.ws.rs.*;
 import jakarta.ws.rs.core.Context;
 import jakarta.ws.rs.core.MediaType;
+import jakarta.ws.rs.core.Response;
 import jakarta.ws.rs.core.SecurityContext;
 import lombok.AllArgsConstructor;
+import org.apache.http.HttpStatus;
+import org.jboss.resteasy.reactive.RestForm;
+import org.jboss.resteasy.reactive.multipart.FileUpload;
 
 @Authenticated
 @Path("/onboarding")
@@ -95,5 +99,19 @@ public class OnboardingController {
 
                     return Uni.createFrom().nullItem();
                 });
+    }
+
+
+
+    @PUT
+    @Path("/{onboardingId}/complete")
+    @Consumes(MediaType.MULTIPART_FORM_DATA)
+    public Uni<Response> complete(@PathParam(value = "onboardingId") String tokenId,
+                                  @RestForm("contract") FileUpload file) {
+
+        return onboardingService.complete(tokenId, file.uploadedFile().toFile(), file.contentType())
+                .onItem().transform(ignore -> Response
+                        .status(HttpStatus.SC_NO_CONTENT)
+                        .build());
     }
 }
